# Compressed chunk performance improvements: Further improvement on precomputed LSB bits