This PR should improve any tsbs read benchmark that is impacted by compressed chuncks performance.


Example of one of the queries impact:
```
make clean build benchmark BENCHMARK=tsbs-scale100-single-groupby-5-8-1.yml
```

master
```
Run complete after 10000 queries with 64 workers (Overall query rate 430.55 queries/sec):
RedisTimeSeries MAX of 5 metrics,random    8 hosts, random � by 1m:
min:     4.09ms, med:   134.81ms, mean:   148.29ms, max:  362.96ms, stddev:    42.86ms, sum: 1482.9sec, count: 10000
all queries                                                         :
min:     4.09ms, med:   134.81ms, mean:   148.29ms, max:  362.96ms, stddev:    42.86ms, sum: 1482.9sec, count: 10000
wall clock time: 23.244326sec
```

branch
```
Run complete after 10000 queries with 64 workers (Overall query rate 448.26 queries/sec):
RedisTimeSeries MAX of 5 metrics,random    8 hosts, random � by 1m:
min:    27.45ms, med:   128.36ms, mean:   142.44ms, max:  305.41ms, stddev:    40.78ms, sum: 1424.4sec, count: 10000
all queries                                                         :
min:    27.45ms, med:   128.36ms, mean:   142.44ms, max:  305.41ms, stddev:    40.78ms, sum: 1424.4sec, count: 10000
```